+++ src/utils.h
@@ -3,6 +3,6 @@
 #include <string>
 
 int str_to_int(const std::string &s);
-int str_to_float(const std::string &s);
+float str_to_float(const std::string &s);
 
 void replace_str(std::string& str, std::string oldstr, std::string newstr);
